+++ basic-slicer/src/types.rs
@@ -1,14 +1,14 @@
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Point3d {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Point2d {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 #[derive(Clone, Debug)]
